Reorganize `:shared:data:model`
