+++ shared/data/core/src/test/java/com/ivy/data/repository/mapper/TransactionMapperTest.kt
@@ -13,7 +13,7 @@ import com.ivy.data.model.Income
 import com.ivy.data.model.TransactionId
 import com.ivy.data.model.TransactionMetadata
 import com.ivy.data.model.Transfer
-import com.ivy.data.model.common.Value
+import com.ivy.data.model.Value
 import com.ivy.data.model.primitive.AssetCode
 import com.ivy.data.model.primitive.AssetCode.Companion.EUR
 import com.ivy.data.model.primitive.AssetCode.Companion.USD
