+++ shared/data/model-testing/src/main/java/com/ivy/data/model/testing/ArbValue.kt
@@ -3,7 +3,7 @@ package com.ivy.data.model.testing
 import arrow.core.None
 import arrow.core.Option
 import arrow.core.getOrElse
-import com.ivy.data.model.common.Value
+import com.ivy.data.model.Value
 import com.ivy.data.model.primitive.AssetCode
 import com.ivy.data.model.primitive.PositiveDouble
 import io.kotest.property.Arb
