+++ shared/data/model-testing/src/test/java/com/ivy/data/model/testing/ArbTransactionTest.kt
@@ -4,7 +4,7 @@ import arrow.core.Some
 import com.google.testing.junit.testparameterinjector.TestParameter
 import com.google.testing.junit.testparameterinjector.TestParameterInjector
 import com.ivy.data.model.AccountId
-import com.ivy.data.model.common.Value
+import com.ivy.data.model.Value
 import com.ivy.data.model.primitive.AssetCode
 import com.ivy.data.model.primitive.PositiveDouble
 import io.kotest.matchers.shouldBe
