+++ shared/data/model/src/main/kotlin/com/ivy/data/model/Tag.kt
@@ -1,6 +1,5 @@
 package com.ivy.data.model
 
-import com.ivy.data.model.common.Reorderable
 import com.ivy.data.model.primitive.ColorInt
 import com.ivy.data.model.primitive.IconAsset
 import com.ivy.data.model.primitive.NotBlankTrimmedString
