+++ shared/data/model/src/main/kotlin/com/ivy/data/model/Value.kt
@@ -1,4 +1,4 @@
-package com.ivy.data.model.common
+package com.ivy.data.model
 
 import com.ivy.data.model.primitive.AssetCode
 import com.ivy.data.model.primitive.PositiveDouble
