+++ shared/data/model/src/main/kotlin/com/ivy/data/model/primitive/AssetCode.kt
@@ -1,7 +1,7 @@
 package com.ivy.data.model.primitive
 
 import arrow.core.raise.Raise
-import com.ivy.data.model.Exact
+import com.ivy.data.model.exact.Exact
 
 @JvmInline
 value class AssetCode private constructor(val code: String) {
