+++ shared/data/model/src/main/kotlin/com/ivy/data/model/primitive/PositiveInt.kt
@@ -2,7 +2,7 @@ package com.ivy.data.model.primitive
 
 import arrow.core.raise.Raise
 import arrow.core.raise.ensure
-import com.ivy.data.model.Exact
+import com.ivy.data.model.exact.Exact
 
 @JvmInline
 value class PositiveInt private constructor(val value: Int) {
