+++ shared/data/model/src/test/java/com/ivy/data/model/primitive/TransactionTest.kt
@@ -6,7 +6,7 @@ import com.ivy.data.model.Income
 import com.ivy.data.model.TransactionId
 import com.ivy.data.model.TransactionMetadata
 import com.ivy.data.model.Transfer
-import com.ivy.data.model.common.Value
+import com.ivy.data.model.Value
 import com.ivy.data.model.getFromAccount
 import com.ivy.data.model.getFromValue
 import com.ivy.data.model.getToAccount
